+++ mockup/patterns/structure/js/views/addmenu.js
@@ -14,7 +14,7 @@ define([
     initialize: function(options) {
       var self = this;
       ButtonGroup.prototype.initialize.apply(self, [options]);
-      $('body').on('context-info-loaded', function(data) {
+      $('body').on('context-info-loaded', function(event, data) {
         self.$items.empty();
         _.each(data.addButtons, function(item) {
           var view = new ButtonView({